still separating common code from client...
