+++ src/configs/maps/map.json
@@ -1,7 +1,7 @@
 {
     "window": {
-        "height": 15,
-        "width": 15,
+        "height": 11,
+        "width": 11,
         "x": 2,
         "y": 4
     },
