+++ src/engine/PositionedObject.js
@@ -25,6 +25,21 @@ class PositionedObject {
             y: this.y + (this.height * offset_percent_y) / 100,
         };
     }
+
+    /**
+     * Координаты объекта относительно окна отображения (канваса)
+     * @param {int} offset_percent_x Сдвиг относительно верхнего левого угла в процентах от размера объекта
+     * @param {int} offset_percent_y Сдвиг относительно верхнего левого угла в процентах от размера объекта
+     */
+    canvasPosition(offset_percent_x = 0, offset_percent_y = 0) {
+        const win = this.map.window;
+        const pos = this.worldPosition(offset_percent_x, offset_percent_y);
+
+        return {
+            x: pos.x - (win ? win.x : 0),
+            y: pos.y - (win ? win.y : 0),
+        };
+    }
 }
 
 Object.assign(PositionedObject.prototype, EventSourceMixin);
